add dbt docs operator with docs on uploading to s3 (#277)
## Description

<!-- Add a brief but complete description of the change. -->

This PR does a few things related to the `dbt docs generate` command:
- adds the `DbtDocsLocalOperator` and a `callback` argument to the base
operator. This lets a user generate docs and take action with them - to,
for example, upload them to be statically served on S3
- adds a new docs page with an example callback that uploads the docs to
S3
- adds an example DAG in the `dev` directory with a callback that
uploads docs to S3 (if a connection is configured). Otherwise, it copies
the docs to a persistent local directory

## Related Issue(s)

<!-- If this PR closes an issue, you can use a keyword to auto-close.
-->
<!-- i.e. "closes #0000" -->
closes #239

## Breaking Change?

<!-- If this introduces a breaking change, specify that here. -->
No breaking changes.

## Checklist

- [x] I have made corresponding changes to the documentation (if
required)
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>