Extend DBT docs operator to upload S3 and Azure Blob Storage (static website) (#287)
## Description

This PR extend DBT docs operators:
- adds the `DbtDocsS3LocalOperator` and
`DbtDocsAzureStorageLocalOperator` to the base DbtDocsLocalOperator
operator. This lets a user generate docs and also take action with them.
These operators handle the appropriate callback invocation automatically
in the background. The user's task is simply to fill in the parameters
required by the operators, and there is no need for them to write a
callback for task execution.
- adds a modified docs page with an examples
- adds a modified DAG in the `dev` directory

## Related Issue(s)

Closes: #286

## Breaking Change?

No.

## Checklist

- [x] I have made corresponding changes to the documentation (if
required)
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Julian LaNeve <lanevejulian@gmail.com>