PR preview for docs page (#300)
## Description

If a change is detected in the **docs/** directory and a commit or pull
request is made to the main branch, it triggers the Netlify build and
publishing of the documentation. The process is not controlled by GitHub
actions, as is likely the case with astronomer-docs, but rather through
Netlify integration. **The build and publish based on the netlify.toml
file.**

After successful publishing, Netlify comments on the pull request with
the URL where the 'deploy preview' can be viewed.

I tested it in my forked repository:

![netlify_test](https://github.com/astronomer/astronomer-cosmos/assets/18029373/f12bc9b6-187a-466d-bbca-d912edee00b9)
Deploy log:
https://app.netlify.com/sites/comfy-macaron-04ffff/deploys/64779d8995ec800008417867
Deploy preview:
https://deploy-preview-14--comfy-macaron-04ffff.netlify.app/

## Instructions for maintainers
* Configure this repository in Netlify
* Enable deploy preview:
https://docs.netlify.com/site-deploys/deploy-previews/#configure-deploy-previews
* Set Github PR comments:
https://docs.netlify.com/site-deploys/notifications/#github-pull-request-comments
* Everything else is handled by netlify.toml

## Related Issue(s)

Closes #139 

## Breaking Change?

No

## Checklist

- [ ] I have made corresponding changes to the documentation (if
required)
- [ ] I have added tests that prove my fix is effective or that my
feature works. (** I checked it in my forked repository**)

---------

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>