[looking for feedback] refactor profiles for maintainability (#271)
## Description

<!-- Add a brief but complete description of the change. -->

This PR refactors how we deal with the Airflow connection -> dbt profile
translation. Previously, there was no standard, meaning maintaining and
contributing profiles was really difficult. This PR fixes that issue by
ensuring there's a consistent way to contribute profiles: a class that
inherits from `BaseProfileMapping` and defines at least:

- `airflow_connection_type`: the Airflow connection type that this
profile mapping is for.
- `required_fields`: a list of required fields for the profile. This
refers to the field as it's named in the dbt profile.
- `secret_fields`: a list of fields that are secret. These fields will
be passed to dbt as environment variables.
- `airflow_param_mapping`: a dictionary that maps the Airflow connection
fields to the dbt profile fields. The keys are the Airflow connection
fields and the values are the dbt profile fields.
- Optionally, a profile mapping can specify a
`transform_{dbt_field_name}` function for each dbt profile field. This
function will be called on the value of the Airflow connection field
before it is passed to dbt. This is useful for transforming the value of
a field before it is passed to dbt. For example, sometimes host fields
need to be passed to dbt without the http:// prefix.

This PR also lays the groundwork for a few user-requested features:
- supplying your own profile.yml instead of translating Airflow
connections
- supplying additional arguments to the profile that aren't present in
the Airflow connection (e.g. `threads`)
- modifying the target name
(https://github.com/astronomer/astronomer-cosmos/issues/266)
- key-based auth
(https://github.com/astronomer/astronomer-cosmos/issues/267)
- various updates to existing connections and new connection requests
(lots of issues here, not linking)

**For reviewers...**

There was no easy way to avoid how large this PR is, because it required
changing every aspect of how we manage profiles. I recommend reviewing a
few files in particular:
-
https://github.com/astronomer/astronomer-cosmos/blob/refactor-profiles/cosmos/providers/dbt/core/profiles/base.py
- this contains logic shared by all new profile mappings
- each of the profiles in [the profiles
dir](https://github.com/astronomer/astronomer-cosmos/tree/refactor-profiles/cosmos/providers/dbt/core/profiles)
that you care about
- the [new docs
page](https://github.com/astronomer/astronomer-cosmos/blob/refactor-profiles/docs/dbt/connections-profiles.rst)

## Related Issue(s)

<!-- If this PR closes an issue, you can use a keyword to auto-close.
-->
<!-- i.e. "closes #0000" -->
nothing explicitly closed by this PR, but this will help with future
issues quite a bit

## Breaking Change?

<!-- If this introduces a breaking change, specify that here. -->
There is a breaking change, but a workaround is included in the code
with a deprecation warning.

Previously, a user would have to specify a `schema` like so:

```python
basic_cosmos_dag = DbtDag(
    dbt_args={"schema": "public"},
    ...,
)
```

Now, you specify it like so:

```python
basic_cosmos_dag = DbtDag(
    profile_args={"schema": "public"},
    ...,
)
```

## Checklist

- [x] I have made corresponding changes to the documentation (if
required)
- [x] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>