Refactor dbt project parsing and Airflow DAG generation (#360)
This is a major refactor related to how we parse and render `dbt` projects in Airflow. It encourages users to move away from custom Cosmos implementation to parse and filter `dbt` projects. It promotes parsing the dbt project using  `dbt ls` or the `dbt` `manifest` file.

Some of the changes:
* Create a standard class (`DbtToAirflowConverter`) to define the shared interface between `DbtDag`
and `DbtTaskGroup`
* Define data classes to exchange information between the different
steps of the Cosmos (parse dbt project, validate arguments, filter,
create Airflow DAG)
* Split the logic of parsing `dbt` projects from the actual Airflow DAG
generation
* Add support to parse dbt projects using `dbt ls`
* Add support to parse dbt projects using their `manifest.json`

Breaking changes:
* `from cosmos.dag import DbtDag` is now `from cosmos.airflow.dag import
DbtDag` or `from cosmos import DbtDag`
* `from cosmos.dag import DbtTaskGroup` is now `from cosmos.airflow.dag
import DbtTaskGroup` or `from cosmos import DbtTaskGroup`
* `select` and `exclude` arguments, passed to `DbtDag` or `DbtTaskGroup`
used to be dictionaries, now they are lists - so we are compatible with
the representation in dbt

**DISCLAIMER**: This PR was not tested with the following:
* validate with K8s operators
* validate with docker operators

This change relates to ticket: #360