update graph parsing mode when virtualenv is selected (#402)
## Description

When using a virtualenv to run our dbt tasks, we are working under the
hypothesis that dbt is not installed in the machine. However as of now,
if we choose virtualenv as the execution mode, dbt tries to parse the
graph of the dbt tasks with **load_via_dbt_ls()**, while it should point
out to **load_via_custom_parser()** due to the fact aforementioned. we
can accomplish it by removing the virutalenv executor from the _elif_
condition for **load_via_dbt_ls()**, so it will be picked up
automatically by the _else_ condition.

## Related Issue(s)

closes #401

## Breaking Change?

No. it worked fine in the dev sandbox, although I did not have the
opportunity to test it with any image without dbt pre-installed.
Obs: I had to change also the tests for this change

## Checklist

- [X] I have made corresponding changes to the documentation (if
required)
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Julian LaNeve <lanevejulian@gmail.com>