Refactor dbt ls to run using symbolic links
To use the original dbt project path has been a source of issues when that path is Read-Only. This issue was faced in the past when running commands which generate target/ and logs/.
This issue is particularly problematic if we want to run [0m22:38:28  Running with dbt=1.6.0
[0m22:38:28  Encountered an error:
Runtime Error
  No dbt_project.yml found at expected path /Users/tati/Code/astronomer-cosmos/dbt_project.yml
  Verify that each entry within packages.yml (and their transitive dependencies) contains a file named dbt_project.yml
   from the project directory, since dbt 1.6 saves adpaters to dbt_packages, unless specified in the user's dbt_project.yml.
This change aims to solve these issues.

Finally, when running dbt ls in a project which has packages.yml, and those have not been installed, dbt raises 'Compilation Error'. Since there may be other errors raised by DBT in stdout, this PR captures Errors in a more generic way.