Add profile_config arg to DbtKubernetesBaseOperator (#505)
## Description

When creating `DbtTaskGroup`s and `DbtDag`s, a `ProfileConfig` needs to
be provided, since it's a required positional argument of the
`DbtAirflowConverter` class. However, when providing a `profile_config`
in `execution_mode="kubernetes"`, the following Exception is thrown:
```
AirflowException: Invalid arguments were passed to DbtRunKubernetesOperator (task_id: my_first_dbt_model_run). Invalid arguments were:
**kwargs: {'profile_config': ProfileConfig(profile_name='jaffle_shop', target_name='dev', profiles_yml_filepath='jaffle_shop/profiles.yml', profile_mapping=None)}
```

In my humble attempt of a solution, I make the `profile_config` an
optional argument of the `DbtKubernetesBaseOperator` class (mainly to
satisfy the tests ... since its a required argument in the previous part
of the "pipeline", having it being a required argument would make sense
imho) and parse it inside the `build_kube_args` method (similar, but not
identically) to how it is done in the `run_command` method of the
`DbtLocalBaseOperator` class.

This solution fixes the thrown exception and has the added benefit of
giving the user that want to run dbt inside kubernetes the option to
specify `--profile` and `--target`, which is a useful feature when
wanting to schedule the same model in airflow, pointing to dev and prod
targets.

I've tested this solution in my GCC environment.

## Related Issue(s)

This PR resolves the issue in issue #493

## Breaking Change?

This should fix something that is broken atm, as far as I can see.

## Checklist

- [ ] I have made corresponding changes to the documentation (if
required)
- [ ] I have added tests that prove my fix is effective or that my
feature works