Fix running dbt tests that depend on multiple models (support --indirect-selection buildable) (#613)
When running tests with dbt, dbt allows users to run them in different
[modes](https://docs.getdbt.com/reference/node-selection/test-selection-examples#indirect-selection).
As described in the linked dbt documentation:
> The "buildable", "cautious", and "empty" modes can be useful in
environments when you're only building a subset of your DAG, and you
want to avoid test failures in "eager" mode caused by unbuilt resources.

Exactly this is the case when using cosmos when running tests with
`AFTER_EACH` behaviour. When a tests depends on multiple models, a test
can result in an error, because the dependent model was not run yet.

Example:
with a simple DAG like this `model1 -> model2`, and a test `rc_test` set
up to test row-count being equal between `model1` and `model2` an error
might occur when we run operations in the following order:
1. `dbt run --model model1`
2. `dbt test --model model1` -> this will result in an error, if the row
count has changed between the last and this run.

adding the argument `--indirect-selection buildable` in step 2 will
avoid actually executing the test there, because dbt makes sure to run
the test only if *all* dependent models are upstream. So we can continue
without an error:

3. `dbt run --model model2`
4. `dbt test --model model2 --indirect-selection buildable` -> this will
actually run the test `rc_test` in this step.

By adding `--indirect-selection buildable` we have avoided running the
test `rc_test` twice (in both the `model1` DagGroup and `model2`
DagGroup), running it only once in `model2`.