Fix adding test nodes to DAGs built using `LoadMethod.DBT_MANIFEST` and `LoadMethod.CUSTOM` (#615)
Resolving issues with the DBT_MANIFEST/CUSTOM load methods when the
has_test attribute is not assigned to the node correctly.

## Description

When a tag selector is used, all tests are filtered out because of the
DbtResourceType.TEST node does not have any information about tags. To
bypass this limitation - tags are assigned to tests based on their
parent model.

## Related Issue(s)

Closes: #580

Co-authored-by: edgarasnavickas <edgaras.navickas@kevin.eu>
(cherry picked from commit 58de67e34243ea38024461173034e9b30c6ee700)