Fix example DAG in the README and docs/index (#705)
The example DAG we had declared in our README and index page was
incompatible with the latest versions of Cosmos.

An Astronomer customer reported this, and we didn't realise it because
those code excerpts were not executed.

This PR changes the references to an example DAG run as part of our
integration tests.

As a follow-up, we should try to avoid this redundancy between
`README.rst` and `index.rst` - which I logged as follow up ticket in
#704.

(cherry picked from commit 2617449bbde6c3f4f099737e704fb623938883d3)