Add new parsing method `LoadMode.DBT_LS_FILE` (#733)
Adds a new load method, which is in between manifest and dbt_ls. This aims to have a dbt ls file with the output prepared and only use the
parsing logic of the dbt_ls load method. This should increase performance
compared to using dbt_ls.

However, this method needs one output file for each taskgroup build.