Additions to docs regarding the `DBT_MANIFEST` load mode (#757)
## Description

I thought that there were a few aspects regarding execution modes that
could require more clarification in the docs.

- The parsing methods docs mentions that only `LOCAL` execution mode is
supported for `DBT_LS`, but the reverse was not true (i.e. execution
mode docs made no mention of parsing methods), so I added notes about
that.
- GCC docs suggest using `VIRTUALENV` execution mode, but makes no
mention of the fact that the `DBT_LS` parsing method is not supported in
this execution mode. Naturally, in this case, users should be utilizing
the `DBT_MANIFEST` load mode, but that means that the docs are
incomplete since they don't include a `manifest_path=?` in the
`ProjectConfig`.
- Note that there are also discussions in the Airflow Slack regarding
issues users have had parsing the `DbtDag` in GCC that are fixable via
using a pre-compiled `manifest,json`, e.g.
https://apache-airflow.slack.com/archives/C059CC42E9W/p1696435273519979
    - Also see #520 for more discussion.
- Generally speaking when doing the `DBT_MANIFEST` load method, the
pattern is that you run `dbt deps && dbt compile` as part of your
deployment, and upload your full dbt project including these artifacts.
This deployment approach may be obvious to veteran users of Airflow
and/or dbt, but it may not be obvious to everyone, so I think adding a
couple sentences in `parsing-methods.rst` is beneficial.

## Related Issue(s)

Not explicitly related, but #520 discusses some issues encountered using
the default parsing method. (Specifically, running `dbt deps` from a
blank slate tends to slow everything down a lot.)

Part of my motivation for adding to the docs is to better advertise +
better document this alternate method of parsing the dbt DAG.

## Breaking Change?

n/a

## Checklist

n/a