Refactor common executor constructors with test coverage (#774)
I noticed in #771 that there was a lot of repeated class constructors in
order to add a new execution mode that is common among `local`, `docker`
and `kubernetes` and there is no test coverage for the constructors and
methods in some of the operators.

This PR attempts to make it easier to add new execution operators in the
future.

## Breaking Change?
None

There may be task UI color differences with the kuberentes/docker
operators, since now all of LS/Seed/Run etc. operators across execution
modes have the same task colors.