Fix: ensure operator `execute` method is consistent across all execution base subclasses (#805)
This fixes an issue reported in #804 after the refactor done in
https://github.com/astronomer/astronomer-cosmos/pull/774 where the
`execute` methods for `DbtLocalBaseOperator`, `DbtDockerBaseOperator`,
and `DbtKubernetesBaseOperator` were different.

This PR refactors the `execute` method to the `AbstractDbtBaseOperator`
so it's the same for all of the local, docker and kubernetes inherited
operators, and adds `build_and_run_cmd` as an abstract method since the
implementation is different across the 3 different execution modes.

Closes #804

(cherry picked from commit 9c090a49b1073537934caf11db5c26cd8b311534)