Improve dbt ls parsing resilence to missing tags/config
An Astronomer customer reported an issue when using LoadMode.DBT_LS and Cosmos raising the error:

E               KeyError: 'tags'

cosmos/dbt/graph.py:113: KeyError

while parsing the DAG in Airflow. It is unclear which of the thousand nodes in their project had this behaviour, but it feels a good idea to make Cosmos more resilient to this use-case regardless.